Reimplemented window frame searching through enumeration for the sake of potential improvements in runtime and better error handling.
Flags and app launching have been temporarily removed, but will be returned in a later commit.

Help and Usage, as well as the top-of-file comment now both contain outdated design descriptions and will need to be rewritten after the v2 design is stable.

Bugfixes:
Teams no longer has to be the focused window at script launch for successful execution.
Outlook no longer has to be on the Inbox folder to be targeted successfully.

Known Issues:
Spotify must not be playing any media in order to be targeted successfully.