Reimplemented flags and args
CLI control of resolution between the two original presets from v1 has been re-added through the command-line argument list.
In order to achieve this within the new enumeration approach to locating window frames the dimension list constants had to be made into variables which are updated in main() with the required scaling operations according to the arguments passed, instead of applying those operations to the initial dimensions at the MoveWindow() call.